+++ tools/gfx_tools/compose/__init__.py
@@ -1,0 +1,4 @@
+from .TileEntry import TileEntry
+from .Tilesheet import Tilesheet
+from .Tileset import Tileset
+from .log import log
