+++ tools/gfx_tools/compose/log.py
@@ -1,0 +1,3 @@
+import logging
+
+log = logging.getLogger(__name__)
